web-app/backend/requirements.txt
numpy
flask
pillow
torch